Police the default subnet too, instead of trusting where it came from
`discover_network` and `sweep_network` with no subnet skipped the target
policy entirely: the default was substituted deep inside `Ops`, after every
check had already run. That was safe, but safe by accident -- the
substituted value happens to be the operator's own interface. Nothing said
so, and it would stop being true the moment anything else supplied a
default.

The subnet is now settled before anything is checked or scanned, so the
policy applies on every path and the string that gets checked is the exact
string that gets scanned, rather than a second derivation that could differ
if an interface changed in between.

The visible consequence is deliberate: a machine whose LAN carries public
addresses -- some universities and ISPs still do this -- now gets a refusal
naming its own subnet instead of silently scanning it. That is the honest
outcome. The operator opts in once, knowing why.

`resolve_and_check` takes the default as a parameter purely so a test can
reach the publicly-addressed-interface case, which is otherwise decided by
whatever interface the test host happens to have. Verified the new test
fails against the old behaviour.

The pcap operations move to their own module: adding the above pushed
operations.rs past the size guard, and every item that moved is gated on
the `pcap` feature and is the only part of the tool surface that writes a
file and holds a capture slot -- a seam rather than an arbitrary cut.
